Transition to new database format.
The old database sorted DJ names and gave each an ID by enumerating them sequentially starting from zero.  This caused a problem in which DJs would not have the same ID when new DJs appeared.  The new database uses CRC32 on the DJ name and assigns the checksum as the DJ's ID.