Set an interval between updates rather than a completion cycle.
A strange bug would occur with the cyles in which the mission threads would stop updating after its initial run.  No exceptions were thrown and no deadlocks were detected.  No threads shared a stop event.  The bug could not be identified, but would go away when the time to wait was a constant.